🔧 Chore: Vanila Extract 초기 세팅
