+++ src/app/page.tsx
@@ -1,5 +1,5 @@
 import Image from "next/image";
-import styles from "./page.module.css";
+import styles from "../styles/page.module.css";
 
 export default function Home() {
   return (
