+++ src/types/routes.d.ts
